NN-5122: E2E int test #2
